Desktop/parser /requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
lxml==4.9.3
fake-useragent==1.4.0
